Removing hacky code Rails 3.0 code
There is a known issue with Rails 3.x and above when using
validates_uniqueness_of and an enum scope. This issue was patched when
providing support for 3.0 but then broke again in 3.1 and 3.2. It
doesn't make sense to have a hack that only works in an older version of
Rails.