Fixes an issue with select_options on nil/blank objects
This was a regression from 2.x where the ActiveRecord object was setting
enum options on itself. Because these options were no longer available,
there was no way to determine the associated enum if the value was nil
or blank. This introduces a sort of null object that is just a subclass
of the base enum that behaves like a nil or blank object.