Declaring enum name is not really a special case
It is common to have multiple enums in an application that do not
share the same name as the field they are representing. There was some
confusion caused by the README because it called this a "special case",
but it's actually quite common. Hopefully this clears that up a bit.

Closes #30