Removes null object from ClassEnum::Base.build
The null object was originally added to retain backwards compatibility
with existing Formtastic support. It was a pain to maintain support for
Formtastic, so support was removed, but the null object remained. The
null object breaks expected behavior if you're expecting the value to be
nil since it does not evaluate to a falsy value.

Previously, blank values had the same methods as enums:

  alarm = Alarm.new(priority: nil)
  alarm.priority.nil? # => true
  alarm.priority.low? # => false

Now, the value will be nil:

  alarm = Alarm.new(priority: nil)
  alarm.priority.nil? # => true
  alarm.priority.low? # NoMethodError: undefined method `low?' for nil:NilClass
  alarm.priority && alarm.priority.low? # => nil